Extend the live provisioning terminal to container pull/build for every project type
The live provisioning terminal covered composer/wp-cli, git clone, and
native Node.js's npm/pnpm install, but the piece common to every
single container-mode project — regardless of type, including plain
PHP and static sites that have no other provisioning step at all —
was still invisible: pulling images and building/starting the
container stack.

container_lifecycle.rs's pull_images()/run_compose() already captured
their docker/podman compose output line-by-line via stream_stderr
(used to detect progress stages and build an error message on
failure), but never showed any of it live. They now also call the
same operations::emit_provision_line() project_templates.rs already
uses, moved there from project_templates.rs since it's no longer
specific to "provisioning" in the composer/wp-cli sense.

containers::operate_for_provisioning(..., "rebuild") — which calls
into these two functions — runs unconditionally for every new
environment during creation regardless of project type, so this
closes the gap end-to-end: image download and container
creation/startup now show up in the same wizard terminal panel for
every container-mode project, not just the ones with their own
composer/git step.
